recruitment-ai-backend/requirements.txt
fastapi
uvicorn
groq
langchain
langchain-community
langchain-huggingface
networkx
chromadb
pypdf
python-multipart
pydantic-settings
python-dotenv